+++ apps/api/Tasks/SetTasksOrder.ts
@@ -2,11 +2,10 @@ import { SharableTaskList, TaskListOrGroup, User } from "@effect-ts-demo/todo-ty
 import * as T from "@effect-ts/core/Effect"
 import * as Lens from "@effect-ts/monocle/Lens"
 
-import { UserSVC } from "@/services"
-
 import * as TaskContext from "./TaskContext"
 
 import { UUID } from "@effect-ts-demo/core/ext/Schema"
+import { UserSVC } from "@effect-ts-demo/infra/services"
 import { Request, Response } from "@effect-ts-demo/todo-client/Tasks/SetTasksOrder"
 
 const inboxOrder = User.lens["|>"](Lens.prop("inboxOrder"))
