+++ apps/api/errors.ts
@@ -1,7 +1,1 @@
-export class NotFoundError {
-  public readonly _tag = "NotFoundError"
-  public readonly message: string
-  constructor(type: string, id: string) {
-    this.message = `Didn't find ${type}#${id}`
-  }
-}
+export * from "@effect-ts-demo/infra/errors"
