+++ packages/infra/errors.ts
@@ -1,0 +1,7 @@
+export class NotFoundError {
+  public readonly _tag = "NotFoundError"
+  public readonly message: string
+  constructor(type: string, id: string) {
+    this.message = `Didn't find ${type}#${id}`
+  }
+}
