+++ packages/infra/express/morphic/routing.ts
@@ -5,19 +5,18 @@ import {
   ParameterLocation,
   SubSchema,
 } from "@atlas-ts/plutus"
+import { schema } from "@atlas-ts/plutus/Schema"
+import * as EO from "@effect-ts-demo/core/ext/EffectOption"
+import { Void } from "@effect-ts-demo/core/ext/Model"
 import { Has, pipe } from "@effect-ts/core"
 import * as A from "@effect-ts/core/Collections/Immutable/Array"
 import * as T from "@effect-ts/core/Effect"
 import * as O from "@effect-ts/core/Option"
 import * as Ex from "@effect-ts/express"
 
-import { makeRequestHandler, RequestHandler } from "@/requestHandlerSchema"
+import { UserSVC } from "../../services"
 
-import { UserSVC } from "./services"
-
-import * as EO from "@effect-ts-demo/core/ext/EffectOption"
-import * as S from "@effect-ts-demo/core/ext/Schema"
-import * as OpenApi from "@effect-ts-demo/core/ext/Schema/Openapi"
+import { makeRequestHandler, RequestHandler } from "./requestHandler"
 
 type Methods = "GET" | "PUT" | "POST" | "PATCH" | "DELETE"
 
@@ -44,7 +43,7 @@ export interface RouteDescriptor<
     ReqA,
     ResA
   >
-  _tag: "Schema"
+  _tag: "Morphic"
 }
 
 export function makeRouteDescriptor<
@@ -71,7 +70,7 @@ export function makeRouteDescriptor<
     ResA
   >
 ) {
-  return { path, method, handler, _tag: "Schema" } as RouteDescriptor<
+  return { path, method, handler, _tag: "Morphic" } as RouteDescriptor<
     R,
     PathA,
     CookieA,
@@ -225,26 +224,20 @@ function del<
 }
 export { del as delete }
 
-export function makeFromSchema(
+export function makeFromMorphic(
   e: RouteDescriptor<any, any, any, any, any, any, any, any>
 ) {
-  const jsonSchema_ = OpenApi.for
-  const jsonSchema = <E, A>(r: S.ReqResSchemed<E, A>) => jsonSchema_(r.Model)
   const { Request: Req, Response: Res } = e.handler
   // TODO: use the path vs body etc serialisation also in the Client.
-  const makeReqQuerySchema = EO.fromNullable(Req.Query)["|>"](
-    EO.chainEffect(jsonSchema)
-  )
+  const makeReqQuerySchema = EO.fromNullable(Req.Query)["|>"](EO.chainEffect(schema))
   const makeReqHeadersSchema = EO.fromNullable(Req.Headers)["|>"](
-    EO.chainEffect(jsonSchema)
+    EO.chainEffect(schema)
   )
-  const makeReqCookieSchema = EO.fromNullable(Req.Cookie)["|>"](
-    EO.chainEffect(jsonSchema)
-  )
-  const makeReqPathSchema = EO.fromNullable(Req.Path)["|>"](EO.chainEffect(jsonSchema))
-  const makeReqBodySchema = EO.fromNullable(Req.Body)["|>"](EO.chainEffect(jsonSchema))
+  const makeReqCookieSchema = EO.fromNullable(Req.Cookie)["|>"](EO.chainEffect(schema))
+  const makeReqPathSchema = EO.fromNullable(Req.Path)["|>"](EO.chainEffect(schema))
+  const makeReqBodySchema = EO.fromNullable(Req.Body)["|>"](EO.chainEffect(schema))
   //const makeReqSchema = schema(Req)
-  const makeResSchema = jsonSchema_(Res)
+  const makeResSchema = schema(Res)
 
   // TODO: custom void type - 204 response
   // https://github.com/Effect-TS/morphic/commit/da3a02fb527089807bcd5253652ee5a5b1efa371
@@ -290,7 +283,7 @@ export function makeFromSchema(
       ),
       responses: A.concat_(
         [
-          e.handler.Response === S.Void
+          e.handler.Response === Void
             ? new Response(204, { description: "Empty" })
             : new Response(200, {
                 description: "OK",
@@ -298,7 +291,7 @@ export function makeFromSchema(
               }),
           new Response(400, { description: "ValidationError" }),
         ],
-        e.path.includes(":") && e.handler.Response === S.Void
+        e.path.includes(":") && e.handler.Response === Void
           ? [new Response(404, { description: "NotFoundError" })]
           : []
       ),
